requirement.txt
apturl==0.5.2
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
Camelot==12.6.29
camelot-py==0.9.0
certifi==2020.6.20
cffi==1.17.1
chardet==4.0.0
charset-normalizer==3.4.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==43.0.1
cupshelpers==1.0
dbus-python==1.2.18
decorator==5.1.1
defer==1.0.6
distlib==0.3.4
distro==1.7.0
distro-info==1.1+ubuntu0.2
docx==0.2.4
duplicity==0.8.21
Elixir==0.7.1
et-xmlfile==1.1.0
exceptiongroup==1.2.2
fasteners==0.14.1
filelock==3.6.0
future==0.18.2
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
iniconfig==2.0.0
jeepney==0.7.1
Jinja2==3.1.4
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lxml==5.3.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
monotonic==1.6
more-itertools==8.10.0
mysql-connector==2.2.9
netifaces==0.11.0
numpy==2.1.2
oauthlib==3.2.0
olefile==0.46
opencv-python==4.10.0.84
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
paramiko==2.9.3
pbr==6.1.0
pdfminer.six==20231228
pdfplumber==0.11.4
pexpect==4.8.0
pillow==10.4.0
platformdirs==2.5.1
pluggy==1.5.0
protobuf==3.12.4
ptyprocess==0.7.0
pycairo==1.20.1
pycparser==2.22
pycups==2.0.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyMuPDF==1.24.11
PyNaCl==1.5.0
pyparsing==2.4.7
pypdf==5.0.1
PyPDF2==3.0.1
pypdfium2==4.30.0
pyRFC3339==1.1
pytest==8.3.3
python-apt==2.4.0+ubuntu3
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-docx==1.1.2
python-pptx==1.0.2
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
SecretStorage==3.3.1
six==1.16.0
SQLAlchemy==0.7.10
sqlalchemy-migrate==0.11.0
sqlparse==0.5.1
systemd-python==234
tabula-py==2.9.3
tabulate==0.9.0
Tempita==0.5.2
tomli==2.0.2
typing_extensions==4.12.2
tzdata==2024.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
virtualenv==20.13.0+ds
wadllib==1.3.6
xdg==5
xkit==0.0.0
xlrd==0.7.1
XlsxWriter==3.2.0
xlwt==0.7.2
zipp==1.0.0
